+++ packages/blade/src/components/Charts/AreaChart/index.ts
@@ -1,0 +1,3 @@
+export { ChartAreaWrapper, ChartArea } from './AreaChart';
+
+export type { ChartAreaWrapperProps, ChartAreaProps } from './types';