+++ packages/blade/src/components/DatePicker/usePopup.ts
@@ -26,7 +26,7 @@ type UsePopupProps = {
   open?: boolean;
   placement: UseFloatingOptions['placement'];
   onOpenChange?: UseFloatingOptions['onOpenChange'];
-  referenceRef: React.RefObject<HTMLButtonElement>;
+  referenceRef: React.RefObject<HTMLButtonElement> | React.RefObject<HTMLElement>;
   crossAxisOffset?: number;
 };