+++ packages/blade/src/components/Input/BaseInput/BaseInput.tsx
@@ -84,7 +84,7 @@ type BaseInputCommonProps = FormInputLabelProps &
     /**
      * Determines if it needs to be rendered as input, textarea or button
      */
-    as?: 'input' | 'textarea' | 'button';
+    as?: 'input' | 'textarea' | 'button' | 'div';
     /**
      * ID that will be used for accessibility
      */
@@ -361,6 +361,11 @@ type BaseInputCommonProps = FormInputLabelProps &
      * @default undefined
      */
     tabIndex?: number;
+    /**
+     * Slot to be rendered adjacent to the value
+     */
+    valueSuffix?: React.ReactNode;
+    children?: ReactNode;
   } & TestID &
   Platform.Select<{
     native: {
@@ -761,33 +766,35 @@ const FocusRingWrapper = styled(BaseBox)<{
   currentInteraction: ActionStates;
   isTableInputCell: NonNullable<BaseInputProps['isTableInputCell']>;
   className: string;
-}>(({ theme, currentInteraction, isTableInputCell }) => ({
+  shouldAddLimitedFocus: boolean;
+}>(({ theme, currentInteraction, isTableInputCell, shouldAddLimitedFocus }) => ({
   borderRadius: makeBorderSize(
     isTableInputCell ? theme.border.radius.none : theme.border.radius.medium,
   ),
   width: '100%',
-  '&:focus-within': !isTableInputCell
-    ? {
-        ...getFocusRingStyles({
-          theme,
-        }),
-        transitionDuration: castWebType(
-          makeMotionTime(
-            getIn(
-              theme.motion.duration,
-              baseInputBorderBackgroundMotion[currentInteraction === 'focus' ? 'enter' : 'exit']
-                .duration,
+  '&:focus-within':
+    !isTableInputCell && (shouldAddLimitedFocus ? currentInteraction === 'focus' : true)
+      ? {
+          ...getFocusRingStyles({
+            theme,
+          }),
+          transitionDuration: castWebType(
+            makeMotionTime(
+              getIn(
+                theme.motion.duration,
+                baseInputBorderBackgroundMotion[currentInteraction === 'focus' ? 'enter' : 'exit']
+                  .duration,
+              ),
             ),
           ),
-        ),
-        transitionTimingFunction: castWebType(
-          theme.motion.easing[
-            baseInputBorderBackgroundMotion[currentInteraction === 'focus' ? 'enter' : 'exit']
-              .easing
-          ],
-        ),
-      }
-    : {},
+          transitionTimingFunction: castWebType(
+            theme.motion.easing[
+              baseInputBorderBackgroundMotion[currentInteraction === 'focus' ? 'enter' : 'exit']
+                .easing
+            ],
+          ),
+        }
+      : {},
 }));
 
 const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps> = (
@@ -864,6 +871,8 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
     trailingDropDown,
     labelSuffix,
     labelTrailing,
+    valueSuffix,
+    children,
     ...rest
   },
   ref,
@@ -962,6 +971,10 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
   }
 
   const isTextArea = as === 'textarea';
+  const hasLeadingDropdown = Boolean(leadingDropDown);
+  const hasTrailingDropdown = Boolean(trailingDropDown);
+
+  const shouldAddLimitedFocus = hasLeadingDropdown || hasTrailingDropdown;
   return (
     <BaseBox
       ref={getOuterMotionRef({ _motionMeta, ref })}
@@ -1004,6 +1017,7 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
           currentInteraction={currentInteraction}
           isTableInputCell={isTableInputCell}
           className="focus-ring-wrapper"
+          shouldAddLimitedFocus={shouldAddLimitedFocus}
         >
           <BaseInputWrapper
             isDropdownTrigger={isDropdownTrigger}
@@ -1061,7 +1075,6 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
               size={_size}
               numberOfLines={numberOfLines}
               isTextArea={isTextArea}
-              hasLeadingDropDown={Boolean(leadingDropDown)}
             >
               <StyledBaseInput
                 as={as}
@@ -1087,6 +1100,7 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
                 trailingInteractionElement={trailingInteractionElement}
                 leadingInteractionElement={leadingInteractionElement}
                 suffix={suffix}
+                valueSuffix={valueSuffix}
                 trailingIcon={trailingIcon}
                 maxCharacters={maxCharacters}
                 textAlign={textAlign}
@@ -1110,6 +1124,7 @@ const _BaseInput: React.ForwardRefRenderFunction<BladeElementRef, BaseInputProps
                 isTableInputCell={isTableInputCell}
                 tabIndex={tabIndex}
                 hasLeadingDropdown={Boolean(leadingDropDown)}
+                children={children}
                 {...metaAttribute({ name: MetaConstants.StyledBaseInput })}
                 {...makeAnalyticsAttribute(rest)}
               />