+++ packages/blade-mcp/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @razorpay/blade-mcp
 
+## 1.7.0
+
+### Minor Changes
+
+- c835336ad: feat(timepicker): added timepicker component
+
+## 1.6.0
+
+### Minor Changes
+
+- ab1773547: feat(blade-mcp): update knowledgebase with AreaChart
+
+## 1.5.0
+
+### Minor Changes
+
+- 2f0e492cd: feat(blade): update knowledge base to support line chart
+
 ## 1.4.5
 
 ### Patch Changes